requirements.txt
kubernetes==31.0.0
PyGithub==2.5.0
PyYAML==6.0.2
requests==2.32.3
python-dateutil==2.9.0
openai==1.57.4
packaging==24.2
